feat(new layout for forms): i redesigned the forms and layout of the details page, not finished yet
