+++ src/App.jsx
@@ -10,7 +10,7 @@ import Spinner from "./utils/Spinner";
 import Home from "./pages/Home";
 import Contact from "./pages/Contact";
 import Templates from "./components/templates/Templates";
-import Details from "./components/details/details";
+import Details from "./pages/Details";
 import Projects from "./components/projects/Projects";
 import Account from "./components/account/Account";
