+++ src/pages/Details.jsx
@@ -1,0 +1,17 @@
+import Header from "../components/header/Header"
+import Footer from "../components/footer/Footer"
+import DetailsContent from "../components/content/DetailsContent"
+
+
+const Details = () => {
+    
+    return (
+        <div className="body">
+            <Header />
+            <DetailsContent />
+            <Footer />
+        </div>
+    )
+}
+
+export default Details
